reverted all the changes and added some documentation to the sortrenderqueue function and a standalone test
