+++ include/MLX42/MLX42_Int.h
@@ -13,10 +13,10 @@
 #ifndef MLX42_INT_H
 # define MLX42_INT_H
 # define LODEPNG_NO_COMPILE_ALLOCATORS
-# include "MLX42.h"
-# include "../lodepng/lodepng.h"
-# include "../glad/glad.h"
-# include "../KHR/khrplatform.h"
+# include "MLX42/MLX42.h"
+# include "lodepng/lodepng.h"
+# include "glad/glad.h"
+# include "KHR/khrplatform.h"
 # if defined(__APPLE__)
 #  define GL_SILENCE_DEPRECATION
 # endif
